requirements.txt
numpy==2.3.1
pandas==2.3.1
plotly==6.0.0
PyYAML==6.0.2
torch==2.7.0+cu128
torch_geometric==2.6.1
matplotlib
seaborn
tqdm
